Update check for split doctor code
