+++ lib/python/qmk/os_helpers/__init__.py
@@ -4,6 +4,7 @@ from enum import Enum
 import re
 import shutil
 import subprocess
+from pathlib import Path
 
 from milc import cli
 from qmk.commands import run
@@ -163,3 +164,21 @@ def check_git_repo():
     dot_git_dir = QMK_FIRMWARE / '.git'
 
     return CheckStatus.OK if dot_git_dir.is_dir() else CheckStatus.WARNING
+
+
+def check_userspace():
+    """Checks if directory configured as external userpace
+
+    If the configured path does not exist or
+    not an absolute path, compilation might fail.
+    """
+    userspace_path = Path(cli.config.user.userspace)
+    if not userspace_path.exists():
+        cli.log.error('External userspace: {fg_red}%s does not exists', userspace_path)
+        return CheckStatus.ERROR
+    elif not userspace_path.is_absolute():
+        cli.log.warning('External userspace: {fg_yellow}%s is not an absolute path', userspace_path)
+        return CheckStatus.WARNING
+    else:
+        cli.log.info('External userspace: {fg_cyan}%s', userspace_path)
+        return CheckStatus.OK
